% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apriori.R
\name{apriori}
\alias{apriori}
\title{Apriori}
\usage{
apriori(X, min_supp = 0.5)
}
\arguments{
\item{X}{Zbior ze zmiennymi binarnymi}

\item{min_supp}{Próg odcięcia zbiorów częstych}
}
\value{
A list
}
\description{
Implementacja algorytmu Apriori dla reguł asocjacyjnych
}
\details{
TODO:...
}
